Seperated scheudle API into its own file
